==> 03_SNV_analysis/01_pairwise_phylogeny.R <==
# From original file:C:\Users\shart\Metzger Lab Dropbox\Sam_data\Mya_genome\revision\BTN_phylogeny.R
# Adapted from original pairwise analysis beginning to output: C:\Users\shart\Metzger Lab Dropbox\Sam_data\Mya_genome\SNPs\somatypus_output_pairwise_noLOH.r

library(tidyverse)
library(ape)
library(seqinr)

setwd("/ssd3/Mar_genome_analysis/somatypus/Mar.3.4.6.p1/final_run")


# Load data if running on all SNVs
    # snvs.nr <- read.table("Somatypus_SNVs_final_NR.txt", header=T, check.names=F)
    # snvs.nv <- read.table("Somatypus_SNVs_final_NV.txt", header=T, check.names=F)
    # snvs.metadata <- read.table("Somatypus_SNVs_final_Metadata.txt", header=T)
    # setwd("/ssd3/Mar_genome_analysis/somatypus/Mar.3.4.6.p1/final_run/pairwise")  
# Load data if running on SNVs outside possible LOH regions
    snvs.nr <- read.table("Somatypus_SNVs_final_noLOH2_NR.txt", header=T, check.names=F)
    snvs.nv <- read.table("Somatypus_SNVs_final_noLOH2_NV.txt", header=T, check.names=F)
    snvs.metadata <- read.table("Somatypus_SNVs_final_noLOH2_Metadata.txt", header=T)
    setwd("/ssd3/Mar_genome_analysis/somatypus/Mar.3.4.6.p1/final_run/pairwise/noLOH")

# Define sample names and combine into one table
    samples <- c("H_ref","H_usa","H_pei",
    "C_pei1","C_pei2","C_pei4","C_pei3", # C_pei4 has high host contamination
    "C_usa1","C_usa2","C_usa6","C_usa3","C_usa7","C_usa4","C_usa5", # C_usa6 & 7 have high host contamination
    "C_pei2T","C_pei4T","C_pei3T","C_usa6T","C_usa3T","C_usa7T","C_usa4T","C_usa5T")
    samplesNR <- paste0(samples, "_NR")
    samplesNV <- paste0(samples, "_NV")
    colnames(snvs.nr) <- samplesNR
    colnames(snvs.nv) <- samplesNV
    colnames(indels.nr) <- samplesNR
    colnames(indels.nv) <- samplesNV
    snvs.metadata2 <- snvs.metadata %>%
        select(CHROM,POS,REF,ALT)
    snvs <- cbind(snvs.metadata2,snvs.nv,snvs.nr)

# Calculate average read depth across the genome
    for(sample in samples){
        depth <- mean(snvs[,paste0(sample,"_NR")])
        assign(paste0(sample,"_avg"), depth)
    }
    for(sample in samples){
        print(sample)
        print(get(paste0(sample,"_avg")))
    }


#Set thresholds
    min_cutoff=3
    stringent_depth=6 # was 8
    weak_depth=16

# Not found in any healthy, found in at least one cancer above stringent threshold
    realSNVs <- filter(snvs, !(H_ref_NV > min_cutoff |
                        H_usa_NV > min_cutoff |
                        H_pei_NV > min_cutoff) &
                        (C_pei1_NV > C_pei1_avg/stringent_depth |
                        C_pei2_NV > C_pei2_avg/stringent_depth |
                        C_pei3_NV > C_pei3_avg/stringent_depth |
                        C_usa1_NV > C_usa1_avg/stringent_depth |
                        C_usa2_NV > C_usa2_avg/stringent_depth |
                        C_usa3_NV > C_usa3_avg/stringent_depth |
                        C_usa4_NV > C_usa4_avg/stringent_depth |
                        C_usa5_NV > C_usa5_avg/stringent_depth))


#####################################################################################

setwd("/ssd3/Mar_genome_analysis/revision/phylogeny")
nrow(snvs)
nrow(realSNVs)
head(realSNVs)

# fasta file of sequences to align

    # Reference genome sequence
        write(">genome", file="variants_concat.fasta",append=TRUE)
        sequence <- realSNVs %>%
            mutate(seq = as.character(REF)) %>%
            .$seq %>%
            paste(collapse = '') %>%
            write(file="variants_concat.fasta",append=TRUE)

    # All sample sequences
        samplesBTN=c("C_pei1","C_pei2","C_pei3","C_usa1","C_usa2","C_usa3","C_usa4","C_usa5")
        for(sample in samplesBTN){
            write(paste0(">", sample), file="variants_concat.fasta",append=TRUE)
            min_depth <- get(paste0(sample, "_avg"))/weak_depth
            realSNVs %>%
                mutate(seq = case_when(get(paste0(sample,"_NV")) > min_depth ~ as.character(ALT),
                                        get(paste0(sample,"_NV")) <= min_depth ~ as.character(REF))) %>%
                .$seq %>%
                paste(collapse = '') %>%
                write(file="variants_concat.fasta",append=TRUE)
        }

    # Convert to phylip format file
        read.dna("variants_concat.fasta", format="fasta") %>%
            write.dna("variants_concat.phy", format="interleaved", nbcol=-1,colsep="")




    # Just sample sequences
        samplesBTN=c("C_pei1","C_pei2","C_pei3","C_usa1","C_usa2","C_usa3","C_usa4","C_usa5")
        for(sample in samplesBTN){
            write(paste0(">", sample), file="variants_concat_BTN.fasta",append=TRUE)
            min_depth <- get(paste0(sample, "_avg"))/weak_depth
            realSNVs %>%
                mutate(seq = case_when(get(paste0(sample,"_NV")) > min_depth ~ as.character(ALT),
                                        get(paste0(sample,"_NV")) < min_depth ~ as.character(REF))) %>%
                .$seq %>%
                paste(collapse = '') %>%
                write(file="variants_concat_BTN.fasta",append=TRUE)
        }
      
    # Convert to phylip format file
        read.dna("variants_concat_BTN.fasta", format="fasta") %>%
            write.dna("variants_concat_BTN.phy", format="interleaved", nbcol=-1,colsep="")

####################################################

# https://a-little-book-of-r-for-bioinformatics.readthedocs.io/en/latest/src/chapter5.html

setwd("/ssd3/Mar_genome_analysis/revision/phylogeny")

# Phylogeny building fuctions
    rootedNJtree <- function(alignment, theoutgroup)
    {
        # load the ape and seqinR packages:
        require("ape")
        require("seqinr")
        # define a function for making a tree:
        makemytree <- function(alignmentmat, outgroup=`theoutgroup`)
        {
            alignment <- ape::as.alignment(alignmentmat)
            alignmentbin <- as.DNAbin(alignment)
            mydist <- dist.dna(alignmentbin, model = "raw")
            mytree <- nj(mydist)
            mytree <- makeLabel(mytree, space="") # get rid of spaces in tip names.
            myrootedtree <- root(mytree, outgroup, r=TRUE)
            return(myrootedtree)
        }
        # infer a tree
        mymat  <- as.matrix.alignment(alignment)
        myrootedtree <- makemytree(mymat, outgroup=theoutgroup)
        # bootstrap the tree
        myboot <- boot.phylo(myrootedtree, mymat, makemytree)
        # plot the tree:
        plot.phylo(myrootedtree, type="p")  # plot the rooted phylogenetic tree
        nodelabels(myboot,cex=0.7)          # plot the bootstrap values
        myrootedtree$node.label <- myboot   # make the bootstrap values be the node labels
        return(myrootedtree)
    }
    unrootedNJtree <- function(alignment)
    {
        # this function requires the ape and seqinR packages:
        require("ape")
        require("seqinr")
        # define a function for making a tree:
        makemytree <- function(alignmentmat)
        {
            alignment <- ape::as.alignment(alignmentmat)
            alignmentbin <- as.DNAbin(alignment)
            mydist <- dist.dna(alignmentbin)
            mytree <- nj(mydist)
            mytree <- makeLabel(mytree, space="") # get rid of spaces in tip names.
            return(mytree)
        }
        # infer a tree
        mymat  <- as.matrix.alignment(alignment)
        mytree <- makemytree(mymat)
        # bootstrap the tree
        myboot <- boot.phylo(mytree, mymat, makemytree)
        # plot the tree:
        plot.phylo(mytree,type="u")   # plot the unrooted phylogenetic tree
        nodelabels(myboot,cex=0.7)    # plot the bootstrap values
        mytree$node.label <- myboot   # make the bootstrap values be the node labels
        return(mytree)
    }


# Run alignments
    loaddata <- read.alignment(file = "variants_concat_BTN.fasta", format = "fasta")
    loaddata2 <- read.alignment(file = "variants_concat.fasta", format = "fasta")

    pdf("nj.tree3.pdf")
    #datatree <- unrootedNJtree(loaddata)
    datatree <- rootedNJtree(loaddata2, "genome")
    dev.off()
    saveRDS(datatree, file="BTN.tree.phy.rds")
    write.tree(datatree, file = "BTN.noLOH.coorectboots.tree")

    pdf("nj.treeBTN.pdf")
    datatree <- unrootedNJtree(loaddata)
    #datatree <- rootedNJtree(loaddata2, "genome")
    dev.off()



#################################################################################
# From original file: C:\Users\shart\Metzger Lab Dropbox\Sam_data\Mya_genome\SNPs\somatypus_output_initial_counts.r

#BINNING - same as for signature extraction
    # Is SNV found in any cancer
        anyC <- filter(snvs, C_pei1_NV > C_pei1_avg/stringent_depth |
                            C_pei2_NV > C_pei2_avg/stringent_depth |
                            C_pei3_NV > C_pei3_avg/stringent_depth |
                            C_usa1_NV > C_usa1_avg/stringent_depth |
                            C_usa2_NV > C_usa2_avg/stringent_depth |
                            C_usa3_NV > C_usa3_avg/stringent_depth |
                            C_usa4_NV > C_usa4_avg/stringent_depth |
                            C_usa5_NV > C_usa5_avg/stringent_depth)
        noC <- filter(snvs, !(C_pei1_NV > C_pei1_avg/stringent_depth |
                            C_pei2_NV > C_pei2_avg/stringent_depth |
                            C_pei3_NV > C_pei3_avg/stringent_depth |
                            C_usa1_NV > C_usa1_avg/stringent_depth |
                            C_usa2_NV > C_usa2_avg/stringent_depth |
                            C_usa3_NV > C_usa3_avg/stringent_depth |
                            C_usa4_NV > C_usa4_avg/stringent_depth |
                            C_usa5_NV > C_usa5_avg/stringent_depth))

    # Is SNV found in any healthy clam    
        anyHanyC <- filter(anyC, H_ref_NV > min_cutoff |
                            H_usa_NV > min_cutoff |
                            H_pei_NV > min_cutoff)
        noHanyC <- filter(anyC, !(H_ref_NV > min_cutoff |
                            H_usa_NV > min_cutoff |
                            H_pei_NV > min_cutoff))
        anyHnoC <- filter(noC, H_ref_NV > min_cutoff |
                            H_usa_NV > min_cutoff |
                            H_pei_NV > min_cutoff)
        noHnoC <- filter(noC, !(H_ref_NV > min_cutoff |
                            H_usa_NV > min_cutoff |
                            H_pei_NV > min_cutoff))
    # which healthy clam is it found in
        anyH <- rbind(anyHnoC, anyHanyC)  %>%
            arrange(CHROM,POS)
        H_R <- filter(noC, (H_ref_NV > min_cutoff) &
                            !(H_usa_NV > min_cutoff) &
                            !(H_pei_NV > min_cutoff))
        H_RU <- filter(noC, (H_ref_NV > min_cutoff) &
                            (H_usa_NV > min_cutoff) &
                            !(H_pei_NV > min_cutoff))
        H_RP <- filter(noC, (H_ref_NV > min_cutoff) &
                            !(H_usa_NV > min_cutoff) &
                            (H_pei_NV > min_cutoff))
        H_RUP <- filter(noC, (H_ref_NV > min_cutoff) &
                            (H_usa_NV > min_cutoff) &
                            (H_pei_NV > min_cutoff))
        H_U <- filter(noC, !(H_ref_NV > min_cutoff) &
                            (H_usa_NV > min_cutoff) &
                            !(H_pei_NV > min_cutoff))
        H_UP <- filter(noC, !(H_ref_NV > min_cutoff) &
                            (H_usa_NV > min_cutoff) &
                            (H_pei_NV > min_cutoff))      
        H_P <- filter(noC, !(H_ref_NV > min_cutoff) &
                            !(H_usa_NV > min_cutoff) &
                            (H_pei_NV > min_cutoff))                                            
    # Likely founder germline
        allCanyH <- filter(anyHanyC, C_pei1_NV > C_pei1_avg/weak_depth &
                            C_pei2_NV > C_pei2_avg/weak_depth &
                            C_pei3_NV > C_pei3_avg/weak_depth &
                            C_usa1_NV > C_usa1_avg/weak_depth &
                            C_usa2_NV > C_usa2_avg/weak_depth &
                            C_usa3_NV > C_usa3_avg/weak_depth &
                            C_usa4_NV > C_usa4_avg/weak_depth &
                            C_usa5_NV > C_usa5_avg/weak_depth)
    # Likely founder germline in LOH region or mismap
        subsetCanyH <- filter(anyHanyC, !(C_pei1_NV > C_pei1_avg/weak_depth &
                            C_pei2_NV > C_pei2_avg/weak_depth &
                            C_pei3_NV > C_pei3_avg/weak_depth &
                            C_usa1_NV > C_usa1_avg/weak_depth &
                            C_usa2_NV > C_usa2_avg/weak_depth &
                            C_usa3_NV > C_usa3_avg/weak_depth &
                            C_usa4_NV > C_usa4_avg/weak_depth &
                            C_usa5_NV > C_usa5_avg/weak_depth))
    # Founder germline OR early somatic mutation
        allCnoH <- filter(noHanyC, C_pei1_NV > C_pei1_avg/weak_depth &
                            C_pei2_NV > C_pei2_avg/weak_depth &
                            C_pei3_NV > C_pei3_avg/weak_depth &
                            C_usa1_NV > C_usa1_avg/weak_depth &
                            C_usa2_NV > C_usa2_avg/weak_depth &
                            C_usa3_NV > C_usa3_avg/weak_depth &
                            C_usa4_NV > C_usa4_avg/weak_depth &
                            C_usa5_NV > C_usa5_avg/weak_depth)
    # Somtic mutations OR mismaps, maybe some founder germline that dropped out of some samples
        subsetCnoH <- filter(noHanyC, !(C_pei1_NV > C_pei1_avg/weak_depth &
                            C_pei2_NV > C_pei2_avg/weak_depth &
                            C_pei3_NV > C_pei3_avg/weak_depth &
                            C_usa1_NV > C_usa1_avg/weak_depth &
                            C_usa2_NV > C_usa2_avg/weak_depth &
                            C_usa3_NV > C_usa3_avg/weak_depth &
                            C_usa4_NV > C_usa4_avg/weak_depth &
                            C_usa5_NV > C_usa5_avg/weak_depth))
    # Sublineage specific
        anyUSAnoPEInoH <- filter(subsetCnoH, !(C_pei1_NV > C_pei1_avg/weak_depth |
                            C_pei2_NV > C_pei2_avg/weak_depth |
                            C_pei3_NV > C_pei3_avg/weak_depth) &
                            (C_usa1_NV > C_usa1_avg/weak_depth |
                            C_usa2_NV > C_usa2_avg/weak_depth |
                            C_usa3_NV > C_usa3_avg/weak_depth |
                            C_usa4_NV > C_usa4_avg/weak_depth |
                            C_usa5_NV > C_usa5_avg/weak_depth))
        anyPEInoUSAnoH <- filter(subsetCnoH, (C_pei1_NV > C_pei1_avg/weak_depth |
                            C_pei2_NV > C_pei2_avg/weak_depth |
                            C_pei3_NV > C_pei3_avg/weak_depth) &
                            !(C_usa1_NV > C_usa1_avg/weak_depth |
                            C_usa2_NV > C_usa2_avg/weak_depth |
                            C_usa3_NV > C_usa3_avg/weak_depth |
                            C_usa4_NV > C_usa4_avg/weak_depth |
                            C_usa5_NV > C_usa5_avg/weak_depth))        
        allUSAnoPEInoH <- filter(subsetCnoH, !(C_pei1_NV > C_pei1_avg/weak_depth |
                            C_pei2_NV > C_pei2_avg/weak_depth |
                            C_pei3_NV > C_pei3_avg/weak_depth) &
                            (C_usa1_NV > C_usa1_avg/weak_depth &
                            C_usa2_NV > C_usa2_avg/weak_depth &
                            C_usa3_NV > C_usa3_avg/weak_depth &
                            C_usa4_NV > C_usa4_avg/weak_depth &
                            C_usa5_NV > C_usa5_avg/weak_depth))
        allPEInoUSAnoH <- filter(subsetCnoH, (C_pei1_NV > C_pei1_avg/weak_depth &
                            C_pei2_NV > C_pei2_avg/weak_depth &
                            C_pei3_NV > C_pei3_avg/weak_depth) &
                            !(C_usa1_NV > C_usa1_avg/weak_depth |
                            C_usa2_NV > C_usa2_avg/weak_depth |
                            C_usa3_NV > C_usa3_avg/weak_depth |
                            C_usa4_NV > C_usa4_avg/weak_depth |
                            C_usa5_NV > C_usa5_avg/weak_depth))

nrow(anyC)
nrow(noC)
nrow(anyHanyC)
nrow(noHanyC)
nrow(anyHnoC)
nrow(noHnoC)
nrow(anyH)
nrow(allCanyH)
nrow(allCnoH)
nrow(anyUSAnoPEInoH)
nrow(anyPEInoUSAnoH)
nrow(allUSAnoPEInoH)
nrow(allPEInoUSAnoH)
    # > nrow(anyC)
    # [1] 10,435,090
    # > nrow(noC)
    # [1] 16,855,311
    # > nrow(anyHanyC)
    # [1] 7,297,123
    # > nrow(noHanyC)
    # [1] 3,137,967
    # > nrow(anyHnoC)
    # [1] 9,002,766
    # > nrow(noHnoC)
    # [1] 7852545
    # > nrow(anyH)
    # [1] 16,299,889
    # > nrow(allCanyH)
    # [1] 6,858,790
    # > nrow(allCnoH)
    # [1] 2,264,329
    # > nrow(anyUSAnoPEInoH)
    # [1] 369,653
    # > nrow(anyPEInoUSAnoH)
    # [1] 330,589
    # > nrow(allUSAnoPEInoH)
    # [1] 331,167
    # > nrow(allPEInoUSAnoH)
    # [1] 320,715
